chore(text-link): move component to UI folder
